tests/wpt/tests/html/semantics/embedded-content/media-elements/networkState_initial.html

<!doctype html>
<html>
 <head>
  <title>{audio,video}.networkState - default state</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <p><a href="https://html.spec.whatwg.org/multipage/#dom-media-networkstate">spec reference</a></p>
  <audio id="a">
  </audio>
  <video id="v">
  </video>
  <div id="log"></div>
  <script>
test(function() {
 var a = document.getElementById("a");
 assert_equals(
  a.networkState,
  a.NETWORK_EMPTY,
  "audioElement.networkState should be NETWORK_EMPTY to begin with");
}, "audio.networkState - default state");

test(function() {
 var v = document.getElementById("v");
 assert_equals(
  v.networkState,
  v.NETWORK_EMPTY,
  "videoElement.networkState should be NETWORK_EMPTY to begin with");
}, "video.networkState - default state");
  </script>
 </body>
</html>
